requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
openai>=1.0.0
python-dotenv>=1.0.0
pypdf>=3.17.0
supabase>=2.0.0
tiktoken>=0.5.0
pymupdf>=1.23.0
pillow>=10.0.0
pytesseract>=0.3.10
llama-index>=0.9.0
llama-index-core>=0.9.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-supabase>=0.1.0
llama-index-readers-file>=0.1.0
